register initial

+++ src/components/shared/Auth/Auth.tsx
@@ -16,27 +16,89 @@ interface AuthProps {
 }
 
 const Auth: FC<AuthProps> = (props) => {
-  const [planets, setPlanets] = useState([]);
-  const [countries, setCountries] = useState([]);
-  const [idenTypes, setIdenTypes] = useState([]);
-  const [isNextPage, setIsNextPage] = useState(false);
+  //Login variables
   const [loginUser, setLoginUser] = useState('');
   const [loginPassword, setLoginPassword] = useState('');
   const { setUser } = useContext(DataContext);
   const navigator = useNavigate();
 
+  //Select tag variables
+  const [planets, setPlanets] = useState([]);
+  const [countries, setCountries] = useState([]);
+  const [idenTypes, setIdenTypes] = useState([]);
+  const [isNextPage, setIsNextPage] = useState(false);
+
+  //Register variables
+  const [email, setEmail] = useState('');
+  const [newUser, setNewUser] = useState('');
+  const [newPassword, setNewPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [name, setName] = useState('');
+  const [lastname, setLastname] = useState('');
+  const [idenType, setIdenType] = useState(0);
+  const [iden, setIden] = useState(0);
+  const [country, setCountry] = useState(0);
+  const [planet, setPlanet] = useState(0);
+
+  //Error variables
+  const [error, setError] = useState('');
+
+  const handleError = (msg: string) => {
+    setError(msg);
+
+    setTimeout(() => setError(''), 2000);
+  };
+
   const handleChange = (el: any, setFunction: any) => setFunction(el);
 
   const loginHandle = async (loginData: any) => {
-    await login({
-      username: loginData.loginUser,
-      password: loginData.loginPassword,
-    }).then(async () => {
-      const userData = await getUserInfo();
-      const result = await userData.json();
-      localStorage.setItem('user', JSON.stringify(result));
-      setUser(result);
-      navigator('/spacebus-ui/');
+    switch (true) {
+      case (!loginUser || loginUser === '') &&
+        (!loginPassword || loginPassword === ''):
+        handleError('Error: Usuario y contraseña vacíos');
+        break;
+
+      case !loginUser || loginUser === '':
+        handleError('Error: Usuario vacío');
+        break;
+
+      case !loginPassword || loginPassword === '':
+        handleError('Error: Contraseña vacía');
+        break;
+
+      default:
+        await login({
+          username: loginData.loginUser,
+          password: loginData.loginPassword,
+        }).then(async (userExists) => {
+          if (userExists) {
+            const userData = await getUserInfo();
+            const result = await userData.json();
+            localStorage.setItem('user', JSON.stringify(result));
+            setUser(result);
+            navigator('/spacebus-ui/');
+          } else {
+            handleError(
+              'Error: Hubo un error de autenticación en el usuario o contraseña'
+            );
+          }
+        });
+        break;
+    }
+  };
+
+  const registerHandle = () => {
+    console.log({
+      email: email,
+      newUser: newUser,
+      newPassword: newPassword,
+      confirmPassword: confirmPassword,
+      name: name,
+      lastname: lastname,
+      idenType: idenType,
+      iden: iden,
+      country: country,
+      planet: planet,
     });
   };
 
@@ -105,6 +167,7 @@ const Auth: FC<AuthProps> = (props) => {
               value={loginPassword}
               onChange={(e) => handleChange(e.target.value, setLoginPassword)}
             />
+            <span className="text-danger mb-2">{error}</span>
             <div className="w-100 text-center">
               <button
                 type="button"
@@ -156,30 +219,32 @@ const Auth: FC<AuthProps> = (props) => {
               <label className={' mb-2'}>Correo</label>
               <input
                 type="email"
-                id="newEmail"
-                name="newEmail"
+                value={email}
                 className={styles.inputField + ' mb-2 border border-light'}
+                onChange={(e) => handleChange(e.target.value, setEmail)}
               />
               <label className={' mb-2'}>Usuario</label>
               <input
                 type="text"
-                id="newUser"
-                name="newUser"
+                value={newUser}
                 className={styles.inputField + ' mb-2 border border-light'}
+                onChange={(e) => handleChange(e.target.value, setNewUser)}
               />
               <label className={' mb-2'}>Contrase&ntilde;a</label>
               <input
                 type="password"
-                id="newPwd"
-                name="newPwd"
+                value={newPassword}
                 className={styles.inputField + ' mb-4 border border-light'}
+                onChange={(e) => handleChange(e.target.value, setNewPassword)}
               />
               <label className={' mb-2'}>Confirmar Contrase&ntilde;a</label>
               <input
                 type="password"
-                id="confirmPwd"
-                name="confirmPwd"
+                value={confirmPassword}
                 className={styles.inputField + ' mb-4 border border-light'}
+                onChange={(e) =>
+                  handleChange(e.target.value, setConfirmPassword)
+                }
               />
 
               <div className="w-100 text-center">
@@ -203,20 +268,23 @@ const Auth: FC<AuthProps> = (props) => {
               <label className={' mb-2'}>Nombre</label>
               <input
                 type="text"
-                id="newName"
-                name="newName"
+                value={name}
                 className={styles.inputField + ' mb-2 border border-light'}
+                onChange={(e) => handleChange(e.target.value, setName)}
               />
               <label className={' mb-2'}>Apellido</label>
               <input
                 type="text"
-                id="newLastname"
-                name="newLastname"
+                value={lastname}
                 className={styles.inputField + ' mb-2 border border-light'}
+                onChange={(e) => handleChange(e.target.value, setLastname)}
               />
               <label className={' mb-2'}>Identificación</label>
               <div className="d-flex w-100 align-items-center justify-content-between">
-                <select className="form-control mb-3 w-25" id="newIdenType">
+                <select
+                  className="form-control mb-3 w-25"
+                  onChange={(e) => handleChange(e.target.value, setIdenType)}
+                >
                   <option value="">Tipo</option>
                   {idenTypes.map((idenType: any) => (
                     <option key={idenType.id} value={idenType.id}>
@@ -227,14 +295,17 @@ const Auth: FC<AuthProps> = (props) => {
 
                 <input
                   type="text"
-                  id="newIden"
-                  name="newIden"
+                  value={iden}
                   className={
                     styles.inputField + ' mb-2 ms-2 border border-light'
                   }
+                  onChange={(e) => handleChange(e.target.value, setIden)}
                 />
               </div>
-              <select className="form-control mb-3 w-25" id="newCountry">
+              <select
+                className="form-control mb-3 w-25"
+                onChange={(e) => handleChange(e.target.value, setCountry)}
+              >
                 <option value="">País</option>
                 {countries.map((country: any) => (
                   <option key={country.id} value={country.id}>
@@ -243,7 +314,10 @@ const Auth: FC<AuthProps> = (props) => {
                 ))}
               </select>
 
-              <select className="form-control mb-3 w-25" id="newFavplanet">
+              <select
+                className="form-control mb-3 w-25"
+                onChange={(e) => handleChange(e.target.value, setPlanet)}
+              >
                 <option value="">Planeta</option>
                 {planets.map((planet: any) => (
                   <option key={planet.id} value={planet.id}>
@@ -252,6 +326,7 @@ const Auth: FC<AuthProps> = (props) => {
                 ))}
               </select>
               <div className="w-100 text-center">
+                <span className="text-danger mb-2">{error}</span>
                 <button
                   type="button"
                   className={styles.button + ' btn btn-info rounded'}
@@ -261,6 +336,7 @@ const Auth: FC<AuthProps> = (props) => {
                 </button>
                 <button
                   className={styles.button + ' btn btn-primary rounded ms-2'}
+                  onClick={() => registerHandle()}
                 >
                   Registrarse
                 </button>
